+++ src/Codebreak.Service.World/Game/Exchange/ShopExchange.cs
@@ -45,7 +45,7 @@ namespace Codebreak.Service.World.Game.Exchange
         /// <param name="actor"></param>
         /// <param name="templateId"></param>
         /// <param name="quantity"></param>
-        public override void BuyItem(EntityBase entity, int templateId, int quantity)
+        public override void BuyItem(EntityBase entity, int templateId, long quantity)
         {
             if (quantity < 1)
             {
